+++ nerfstudio/engine/trainer.py
@@ -21,16 +21,13 @@ import dataclasses
 import functools
 import os
 import time
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import Dict, List, Optional, Tuple, Type
+from typing import Dict, List, Tuple
 
 import torch
 from rich.console import Console
 from torch.cuda.amp.grad_scaler import GradScaler
 from typing_extensions import Literal
 
-from nerfstudio.configs.base_config import InstantiateConfig
 from nerfstudio.configs.experiment_config import ExperimentConfig
 from nerfstudio.engine.callbacks import (
     TrainingCallback,
@@ -50,37 +47,6 @@ from nerfstudio.utils.writer import EventName, TimeWriter
 from nerfstudio.viewer.server import viewer_utils
 
 CONSOLE = Console(width=120)
-
-
-# Trainer related configs
-@dataclass
-class TrainerConfig(InstantiateConfig):
-    """Configuration for training regimen"""
-
-    _target: Type = field(default_factory=lambda: Trainer)
-    """target class to instantiate"""
-    steps_per_save: int = 1000
-    """Number of steps between saves."""
-    steps_per_eval_batch: int = 500
-    """Number of steps between randomly sampled batches of rays."""
-    steps_per_eval_image: int = 500
-    """Number of steps between single eval images."""
-    steps_per_eval_all_images: int = 25000
-    """Number of steps between eval all images."""
-    max_num_iterations: int = 1000000
-    """Maximum number of iterations to run."""
-    mixed_precision: bool = False
-    """Whether or not to use mixed precision for training."""
-    relative_model_dir: Path = Path("nerfstudio_models/")
-    """Relative path to save all checkpoints."""
-    save_only_latest_checkpoint: bool = True
-    """Whether to only save the latest checkpoint or all checkpoints."""
-    # optional parameters if we want to resume training
-    load_dir: Optional[Path] = None
-    """Optionally specify a pre-trained model directory to load from."""
-    load_step: Optional[int] = None
-    """Optionally specify model step to load from; if none, will find most recent model in load_dir."""
-    load_config: Optional[Path] = None
 
 
 class Trainer:
